:sparkles: BE_게시글 최신순 조회, 좋아요 등록/취소, 조회수 기능 업데이트

+++ backend/src/main/java/com/a601/backend/api/service/BoardServiceImpl.java
@@ -1,16 +1,22 @@
 package com.a601.backend.api.service;
 
+import com.a601.backend.api.domain.dto.request.BoardLikeRequest;
 import com.a601.backend.api.domain.dto.request.BoardRequest;
+import com.a601.backend.api.domain.dto.request.UserRequest;
 import com.a601.backend.api.domain.dto.response.BoardResponse;
 import com.a601.backend.api.domain.entity.Board;
+import com.a601.backend.api.domain.entity.BoardLike;
 import com.a601.backend.api.domain.entity.User;
 import com.a601.backend.api.domain.enums.ErrorCode;
 import com.a601.backend.api.exception.CustomException;
+import com.a601.backend.api.repository.BoardLikeRepository;
 import com.a601.backend.api.repository.BoardRepository;
 import com.a601.backend.api.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -23,6 +29,8 @@ public class BoardServiceImpl implements BoardService {
     private final BoardRepository boardRepository;
     private final UserRepository userRepository;
 
+    private final BoardLikeRepository boardLikeRepository;
+
 
     @Override
     public BoardResponse findByBoardId(Long boardId) {
@@ -33,9 +41,11 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public Page<BoardResponse> findAll(Pageable pageable) {
-        return boardRepository.findAll(pageable).map(BoardResponse::new);
+    public Page<BoardResponse> findAllByOrderByCreatedTime(Pageable pageable) {
+        return boardRepository.findAllByOrderByCreatedTimeDesc(pageable).map(BoardResponse::new);
     }
+
+
 
     @Override
     public Long writeBoard(BoardRequest request) {
@@ -50,7 +60,10 @@ public class BoardServiceImpl implements BoardService {
 
     @Override
     public void deleteBoard(Long boardId) {
-        boardRepository.deleteById(boardId);
+        Board board = boardRepository.findById(boardId).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND));
+        if(board != null){
+            boardRepository.deleteById(boardId);
+        }
     }
 
     @Override
@@ -63,9 +76,45 @@ public class BoardServiceImpl implements BoardService {
     @Override
     @Transactional
     public void updateHit(Long boardId) {
-        Board board = boardRepository.findById(boardId).orElseThrow(()-> new CustomException(ErrorCode.POSTS_NOT_FOUND));
-        board.increaseHit();
 
+        boardRepository.findById(boardId).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND)).increaseHit();
+
+    }
+
+    @Override
+    public Page<BoardResponse> findByTitleContaining(String keyword, Pageable pageable) {
+
+        return boardRepository.findByTitleContainingOrderByCreatedTimeDesc(keyword, pageable).map(BoardResponse::new);
+    }
+
+    @Override
+    @Transactional
+    public void addBoardLike(BoardLikeRequest boardLikeRequest) {
+        Board board = boardRepository.findById(boardLikeRequest.getBoardId()).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND));
+        User user = userRepository.findById(boardLikeRequest.getUserId()).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND));
+        // 좋아요 1 증가
+        board.addBoardLike();
+        BoardLike boardLike = BoardLike.builder()
+                .user(user)
+                .board(board)
+                .build();
+        
+        // 좋아요 정보 저장
+        boardLikeRepository.save(boardLike);
+
+    }
+
+    @Override
+    @Transactional
+    public void cancelBoardLike(BoardLikeRequest boardLikeRequest) {
+        BoardLike boardLike = boardLikeRepository.findById(boardLikeRequest.getBoardLikeId()).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND));
+        Board board = boardRepository.findById(boardLikeRequest.getBoardLikeId()).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND));
+        if(boardLike != null){
+            boardLikeRepository.deleteById(boardLike.getBoardLikeId());
+        }
+        if(board != null){
+            board.cancelBoardLike();
+        }
     }
